functions/requirements.txt
firebase-functions==0.1.0
PyJWT==2.8.0
requests==2.31.0
